requirements.txt
beautifulsoup4==4.11.1
googlesearch_python==1.0.1
nmap==0.0.1
PyPDF2==1.27.12
python_nmap==0.7.1
requests==2.25.1
